Whats wrong with Public Video Surveillance??
CCTV or Video surveillance systems are trending more. Ofcourse it protects from theft and crime. Public video surveillance is like a boon and also a bane.  Some  cheaper cameras also arriving in the society. The public video surveillance has many valuable good things and also some drawbacks.  As “good and bad are the two sides of a coin”, its one of such.  Video surveillance can be used as a tool of blackmail.  Public video surveillance have the limited benefits. They decreased the number of crimes.